Moves custom gear to it's own subaddon
